minor clamav.h (and related) cleanup
git-svn-id: file:///var/lib/svn/clamav-devel/trunk/clamav-devel@2470 77e5149b-7576-45b1-b177-96237e5ba77b

+++ libclamav/cvd.c
@@ -217,7 +217,7 @@ struct cl_cvd *cl_cvdparse(const char *head)
 	free(cvd);
 	return NULL;
     }
-    cvd->fl = (short int) atoi(pt);
+    cvd->fl = atoi(pt);
     free(pt);
 
     if(!(cvd->md5 = cli_strtok(head, 5, ":"))) {
